sim: Add a Drain interface that refreshes memory buffers
Add the method Drainable::memRefresh() that functionally updates the
contents of memory buffers with new from main memory. The default
implementation calls memInvalidate() which makes sure that components
not supporting this method still behave in a functionally correct
way. The intention of this method is to allow some contents to be
store in a buffer (e.g., a cache) when switching between a virtualized
and a simulated CPU. The control script would in this case write back
any dirty data (m5.memWriteback(root)) prior to switching to the
virtualized CPU and call m5.memRefresh(root) to update stale data
while switching back.