mem: Add support for huge pages in backing store
Add support for using huge pages for the backing store when
available. This feature automatically tries to discover support for
huge pages at runtime (using the /sys file system). If allocations
using huge pages fail, they automatically fall back to normal pages.

This is the initial version of this patch. It has the following known
limitations:

 * Does not test for MAP_HUGETLB at compile time. Fails when compiled
   with old kernel headers.

 * Does not work on OSX.

 * Checkpointing is untested.