sim: Add a curTick() method to the EventManager class
Add a curTick() method to EventManager classes. This ensures that
SimObjects see the time of their event queues rather than their the
time of the currently executing thread/eq. This update also updates
objects to explicitly request the curTick() from the right event
queue. To simplify a common case of scheduling events, this changeset
adds the method EventManager::scheduleRelative() to simplify the
common case of scheduling an event N ticks into the future.