kvm: Lock the event queue runing the VM when doing IO
In order to support multiple CPUs running in the same VM, we need to
lock the event queue simulating devices when performing an IO. This
ensures that only one CPU ever accesses devices at a time. We
implement this by making the VM support the basic locking protocol by
forwarding it to its event queue. This implies that VMs and devices
must live in the same event queue.