kvm: x86: Always assume segments to be usable
When transfering segment registers into kvm, we need to find the value
of the unusable bit. We used to assume that this can be inferred from
the selector since segments are generally unusable if their selector
is 0. This doesn't work in practice due to some weird corner cases.
Instead, just assume that segments are always usable. This is what
qemu does so it should work.