x86: Set CS limit to 0xffff on INIT
It seems like gem5 used to store segments limits as the actual limit
in memory (x86 stores the segment size). However, the ISA code doesn't
seem to do the necessary calculations, so in practice, we seem to
store the size of the segment rather than size+base. This changeset
changes the CS limit to exclude the base of the segment when
initialized by the INIT interrupt.