kvm: Add support for multicore simulation
Simulating a SMP or multicore requires devices to be shared across KVM
vCPUs. This requires locking when accessing devices. This changeset
adds the neccessary locking to allow devices to execute correctly.
This is implemented by migrating the temporarily migrating the KVM CPU
to the VM's (and devices) event queue when handling MMIO. Similarly,
the VM migrates to the interrupt controller's event queue when
delivering an interrupt.