x86: Allow interrupt delivery across threads
Make it possible to have the interrupt controller and the other device
models in different threads. This changeset makes the interrupt
controller temporarily migrate to its "home thread" when delivering an
interrupt or handling a MMIO request. This makes it possible to use
configurations where the interrupt controller and a CPU live in one
thread and device models in another.